backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
Flask-MySQLdb==1.0.1
PyMySQL==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
email-validator==2.0.0